Refactoring and making workers able to do  multiple jobs in a row
Reworking uint16 -> uint32 when working with GridIndexes to enable larger maps.
Added function to generate random maps.
Added function to print out results from workers

+++ AStar/AStarDefs.h
@@ -4,5 +4,5 @@
 
 namespace NAStarDefs
 {
-	constexpr uint8 NUM_OF_NEIGHBOURS = 4;
+	constexpr uint32 NUM_OF_NEIGHBOURS = 4;
 }
